+++ examples/i18n/src/i18n.rs
@@ -1,24 +1,20 @@
 use fluent::{FluentArgs, FluentBundle, FluentResource};
-use strum_macros::{AsRefStr, EnumIter, EnumString};
+use strum_macros::{AsRefStr, Display, EnumIter, EnumString};
 use unic_langid::LanguageIdentifier;
-
-use super::resource::Resource;
 
 // ------ I18n ------
 
 pub struct I18n {
     lang: Lang,
-    resource: FluentBundle<FluentResource>,
+    ftl_bundle: FluentBundle<FluentResource>,
 }
 
 impl I18n {
     pub fn new(lang: Lang) -> Self {
-        let mut i18n = Self {
+        Self {
             lang,
-            resource: FluentBundle::default(),
-        };
-        i18n.set_lang(lang);
-        i18n
+            ftl_bundle: lang.create_ftl_bundle(),
+        }
     }
 
     pub const fn lang(&self) -> &Lang {
@@ -27,29 +23,19 @@ impl I18n {
 
     pub fn set_lang(&mut self, lang: Lang) -> &Self {
         self.lang = lang;
-        let ftl_res = FluentResource::try_new(
-            Resource::new()
-                .get(lang.as_ref())
-                .expect("get language identifier")
-                .parse()
-                .expect("parse language identifier"),
-        )
-        .expect("parse FTL string");
-
-        let locale: LanguageIdentifier = lang.as_ref().parse().expect("parse language identifier");
-
-        let mut bundle = FluentBundle::new(&[locale]);
-        bundle.add_resource(ftl_res).expect("add FTL resource");
-
-        self.resource = bundle;
+        self.ftl_bundle = lang.create_ftl_bundle();
         self
     }
 
-    pub fn translate(&self, key: &str, args: Option<&FluentArgs>) -> String {
-        let fluent_msg = self.resource.get_message(key).expect("get fluent message");
-        let pattern = fluent_msg.value.expect("get value for fluent message");
+    pub fn translate(&self, key: impl AsRef<str>, args: Option<&FluentArgs>) -> String {
+        let mmessage = self
+            .ftl_bundle
+            .get_message(key.as_ref())
+            .expect("get fluent message");
 
-        self.resource
+        let pattern = mmessage.value.expect("get value for fluent message");
+
+        self.ftl_bundle
             .format_pattern(pattern, args, &mut vec![])
             .to_string()
     }
@@ -57,8 +43,7 @@ impl I18n {
 
 // ------ Lang ------
 
-#[allow(non_camel_case_types)]
-#[derive(Copy, Clone, EnumIter, EnumString, AsRefStr, PartialEq)]
+#[derive(Debug, Copy, Clone, Display, EnumIter, EnumString, AsRefStr, Eq, PartialEq)]
 pub enum Lang {
     #[strum(serialize = "en-US")]
     EnUS,
@@ -67,10 +52,80 @@ pub enum Lang {
 }
 
 impl Lang {
-    pub fn label(&self) -> &str {
+    pub fn label(self) -> &'static str {
         match self {
             Self::EnUS => "English (US)",
             Self::DeDE => "Deutsch (Deutschland)",
         }
     }
+
+    pub fn ftl_messages(self) -> &'static str {
+        macro_rules! include_ftl_messages {
+            ( $lang_id:literal ) => {
+                include_str!(concat!("../ftl_messages/", $lang_id, ".ftl"))
+            };
+        }
+        match self {
+            Self::EnUS => include_ftl_messages!("en-US"),
+            Self::DeDE => include_ftl_messages!("de-DE"),
+        }
+    }
+
+    pub fn to_language_identifier(self) -> LanguageIdentifier {
+        self.as_ref()
+            .parse()
+            .expect("parse Lang to LanguageIdentifier")
+    }
+
+    pub fn create_ftl_bundle(self) -> FluentBundle<FluentResource> {
+        let ftl_resource =
+            FluentResource::try_new(self.ftl_messages().to_owned()).expect("parse FTL messages");
+
+        let mut bundle = FluentBundle::new(&[self.to_language_identifier()]);
+        bundle.add_resource(ftl_resource).expect("add FTL resource");
+        bundle
+    }
+}
+
+// ------ create_t ------
+
+/// Convenience macro to improve readability of `view`s with many translations.
+///
+/// # Example
+///
+///```rust,no_run
+/// fn view(model: &Model) -> Node<Msg> {
+///    let args_male_sg = fluent_args![
+///      "userName" => "Stephan",
+///      "userGender" => "male",
+///    ];
+///
+///    create_t!(model.i18n);
+///    div![
+///        p![t!("hello-world")],
+///        p![t!("hello-user", args_male_sg)],
+///    ]
+/// }
+///```
+#[macro_export]
+macro_rules! create_t {
+    ( $i18n:expr ) => {
+        // This replaces $d with $ in the inner macro.
+        seed::with_dollar_sign! {
+            ($d:tt) => {
+                macro_rules! t {
+                    { $d key:expr } => {
+                        {
+                            $i18n.translate($d key, None)
+                        }
+                    };
+                    { $d key:expr, $d args:expr } => {
+                        {
+                            $i18n.translate($d key, Some(&$d args))
+                        }
+                    };
+                }
+            }
+        }
+    };
 }